fix all bugs on users

+++ server/users/models/usersAccessDataService.js
@@ -47,37 +47,40 @@ const loginUser = async ({ email, password }) => {
 const getUsers = async () => {
   if (DB === "MONGODB") {
     try {
-      const users = User.find({}, { password: 0, __v: 0 });
-
-      //   throw new Error("Opss... i did it again!");
+      const users = await User.find({}, { password: 0, __v: 0 });
       return Promise.resolve(users);
     } catch (error) {
-      return handleBadRequest("Mongoose", error);
+      error.status = 404;
+      return Promise.reject(error);
     }
   }
-  return Promise.resolve({});
+  return Promise.resolve("get users not in mongodb");
 };
 
 const getUser = async (userId) => {
   if (DB === "MONGODB") {
     try {
-      const user = User.findById({ userId }, { password: 0, __v: 0 });
-
-      //   throw new Error("Opss... i did it again!");
+      let user = await User.findById(userId, {
+        password: 0,
+        isAdmin:0,
+        __v: 0,
+      });
+      if (!user) throw new Error("Could not find this user in the database");
       return Promise.resolve(user);
     } catch (error) {
+      error.status = 404;
       return handleBadRequest("Mongoose", error);
     }
   }
   return Promise.resolve("get user not in mongodb");
 };
 
-const updateUser = async (userId, normalizedUser) => {
+const updateUser =  async (id, normalizeUser) => {
   if (DB === "MONGODB") {
     try {
-      const user = User.findByIdAndUpdate(userId, normalizedUser, {
+      let user = await User.findByIdAndUpdate( id, await normalizeUser, {
         new: true,
-      });
+      }).select(["-password", "-__v"]);
 
       if (!user)
         throw new Error(
@@ -85,31 +88,39 @@ const updateUser = async (userId, normalizedUser) => {
         );
       return Promise.resolve(user);
     } catch (error) {
+      error.status = 404;
       return handleBadRequest("Mongoose", error);
     }
   }
-  return Promise.resolve("card update not in mongodb");
+  return Promise.resolve("User Updated!");
 };
 
-const changeUserBusinessStatus = async (userId) => {
+const changeUserBusinessStatus = async id => {
   if (DB === "MONGODB") {
     try {
       const pipeline = [{ $set: { isBusiness: { $not: "$isBusiness" } } }];
       const user = await User.findByIdAndUpdate(id, pipeline, {
         new: true,
-      });
+      }).select(["-password", "-__v", "-isAdmin"]);
+
+      if (!user)
+        throw new Error(
+          "Could not update this user isBusiness status because a user with this ID cannot be found in the database"
+        );
       return Promise.resolve(user);
     } catch (error) {
+      error.status = 404;
       return handleBadRequest("Mongoose", error);
     }
   }
-  return Promise.resolve("card updated!");
+
+  return Promise.resolve("Card Updated!");
 };
 
-const deleteUser = async (userId) => {
+const deleteUser = async (id) => {
   if (DB === "MONGODB") {
     try {
-      let user = await User.findByIdAndDelete(id, { password: 0, __v: 0 });
+      let user = await User.findByIdAndDelete(id, { password: 0,isAdmin:0, __v: 0 });
       if (!user)
         throw new Error(
           "Could not delete this user because a user with this ID cannot be found in the database"
